Final refactoring

+++ src/by/epamtc/exercise3/unit/GeometryLogics.java
@@ -2,15 +2,15 @@ package by.epamtc.exercise3.unit;
 
 public class GeometryLogics {
 
-    public static int countCircleInSquareRadius(int area) {
-        return (int) Math.sqrt(area) / 2;
+    public static double countCircleInSquareRadius(double area) {
+        return Math.sqrt(area) / 2;
     }
 
-    public static int countSquareInCircleArea(int radius) {
-        return (int) (2 * Math.pow(radius, 2));
+    public static double countSquareInCircleArea(double radius) {
+        return 2 * Math.pow(radius, 2);
     }
 
-    public static int countSquareDifference(int square1, int square2) {
+    public static double countSquareDifference(double square1, double square2) {
         return square1 / square2;
     }
